feat: Community Section으로 관리
