+++ src/containers/CommunityPostListSection/CommunityPost/CommunityPost.style.ts
